e2e tests: move the test configuration to a pkg
